Tweaked routines for using facebook JS API
This webapp only needs userID, username, email and profilepic from FB.
Therefore, I have made some changes in order to only make API calls for the first time the website is opened.
It checks the session, and if the info is not there, it tries to authenticate. If the user has allowed access for this APP to his profile info at some point, it will be automatically saved in session. If not, the user can log in the messages page.

These info will be used to authorize user to post messages and delete his/her previous messages.
It will also identify the user's session in tomcat.